STAR-1792: fixed osd error

+++ src/ska_oso_services/pht/models/domain.py
@@ -156,7 +156,7 @@ class AA2_LOW(BaseModel):
     number_beams: int
     max_baseline_km: int
     available_bandwidth_hz: int
-    channel_width_hz: int
+    channel_width_hz: float
     cbf_modes: list[str]
     number_zoom_windows: int
     number_zoom_channels: int
